Report success/failure after the last after_each()
Treat the sequence of before_each()/it()/after_each() for each test as a
whole and report success/failure only after all of them have been run.

Exchange the type of the test step and the last after_each step so that
outcome is reported only after the last after_each step has been run.
The steps keeps their full_name, which is used to identify which step to
run, but the last after_each gets the name of the test step, which is
used in reporting.